requirements.txt
Django==3.1.6
psycopg2 --no-binary psycopg2
djangorestframework
pytest-django
